- Update architecture and di for repository
