+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/data/repository/local/tvseries/LocalTvSeriesRepository.kt
@@ -1,17 +1,8 @@
 package com.piashcse.hilt_mvvm_compose_movie.data.repository.local.tvseries
 
-import com.piashcse.hilt_mvvm_compose_movie.data.datasource.local.dao.FavoriteTvSeriesDao
 import com.piashcse.hilt_mvvm_compose_movie.data.model.tv_series_detail.TvSeriesDetail
-import javax.inject.Inject
 
-class LocalTvSeriesRepository @Inject constructor(
-    private val tvSeriesDao: FavoriteTvSeriesDao,
-) : LocalTvSeriesRepositoryInterface {
-    override suspend fun favoriteTvSeries(): List<TvSeriesDetail?> {
-        return tvSeriesDao.getAllTvSeriesDetails()
-    }
-
-    override suspend fun removeTvSeriesById(tvSeriesId: Int) {
-        tvSeriesDao.deleteTvSeriesById(tvSeriesId)
-    }
+interface LocalTvSeriesRepository {
+    suspend fun favoriteTvSeries(): List<TvSeriesDetail?>
+    suspend fun removeTvSeriesById(tvSeriesId: Int)
 }
