--- app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/data/repository/local/tvseries/LocalTvSeriesRepositoryInterface.kt
@@ -1,8 +1,0 @@
-package com.piashcse.hilt_mvvm_compose_movie.data.repository.local.tvseries
-
-import com.piashcse.hilt_mvvm_compose_movie.data.model.tv_series_detail.TvSeriesDetail
-
-interface LocalTvSeriesRepositoryInterface {
-    suspend fun favoriteTvSeries(): List<TvSeriesDetail?>
-    suspend fun removeTvSeriesById(tvSeriesId: Int)
-}
