+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/data/repository/remote/artist/ArtistRepository.kt
@@ -1,34 +1,11 @@
 package com.piashcse.hilt_mvvm_compose_movie.data.repository.remote.artist
 
-import com.piashcse.hilt_mvvm_compose_movie.data.datasource.remote.ApiService
 import com.piashcse.hilt_mvvm_compose_movie.data.model.artist.ArtistDetail
 import com.piashcse.hilt_mvvm_compose_movie.data.model.artist.ArtistMovies
 import com.piashcse.hilt_mvvm_compose_movie.utils.network.DataState
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import javax.inject.Inject
 
-class ArtistRepository @Inject constructor(
-    private val apiService: ApiService,
-) : ArtistRepositoryInterface {
-    override suspend fun artistAllMovies(movieId: Int): Flow<DataState<ArtistMovies>> = flow {
-        emit(DataState.Loading)
-        try {
-            val searchResult = apiService.artistAllMovies(movieId)
-            emit(DataState.Success(searchResult))
-
-        } catch (e: Exception) {
-            emit(DataState.Error(e))
-        }
-    }
-    override suspend fun artistDetail(personId: Int): Flow<DataState<ArtistDetail>> = flow {
-        emit(DataState.Loading)
-        try {
-            val artistDetailResult = apiService.artistDetail(personId)
-            emit(DataState.Success(artistDetailResult))
-
-        } catch (e: Exception) {
-            emit(DataState.Error(e))
-        }
-    }
+interface ArtistRepository {
+    suspend fun artistAllMovies(movieId: Int): Flow<DataState<ArtistMovies>>
+    suspend fun artistDetail(personId: Int): Flow<DataState<ArtistDetail>>
 }
