+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/ui/screens/celebrities/popular/PopularCelebrities.kt
@@ -4,9 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import androidx.paging.compose.collectAsLazyPagingItems
-import com.piashcse.hilt_mvvm_compose_movie.data.model.GenreId
 import com.piashcse.hilt_mvvm_compose_movie.ui.component.Celebrities
-import com.piashcse.hilt_mvvm_compose_movie.ui.component.TvSeries
 
 @Composable
 fun PopularCelebrities(
